SecureInfoRegistry/pyproject.toml
[project]
name = "repl-nix-secureinforegistry"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "twilio>=9.3.4",
    "flask-mail>=0.10.0",
    "cryptography>=43.0.3",
    "pillow>=11.0.0",
    "sqlalchemy",
]
